src/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.9
django-celery-beat==2.5.0
django-timezone-field==6.1.0
flake8==7.0.0
Flask==3.0.0
flower==2.0.1
gunicorn==21.2.0
humanize==4.9.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
loguru==0.7.2
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.2
pika==1.3.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
